Add search functionality in shared notes
A new function `searchSharedNotesByTitleAndPathology` has been added in both `noteSharesValidation.js` and `notesSharesService.js`, which allows to search the shared notes by title and pathology. A new component `DocumentsSearch.js` is added in SharedNotes which provides the UI for this search functionality. This feature is implemented by keeping user-friendly interaction in the forefront so the application user could easily search shared notes. Changes are also reflected in `sharedNotesActions.js`, to handle the state after a successful search.