"Remove redundant modules and update file structure"
Several redundant JavaScript files in the 'components' directory were deleted to eliminate unnecessary code and declutter the project structure. File paths were updated in various other .js files to ensure accurate references. The "my-documents" section was also commented out in the webpack.common.js file as it was not in use. These changes significantly enhance the readability and maintainability of the codebase.