"Add UI loading state on notes transitions"
Added a loading state display during transition between notes by changing the `isRuleStatusLoading` variable of the `StateStore.MY_NOTES` store in both `my-notes` and `my-shared-documents` navigation routes. This gives the user visual feedback that the application is processing their request.

Also, modified `DocumentsDataTable.js` file to correctly show file icons on pagination. We now use `Alpine.effect` to listen for changes in the noteList, and then update the fileIcons of the current note. This fixes a bug where file icons were not being displayed when the user navigated to a different page on the notes list.