Add email validation and alert messages in share modal
The update introduces email field validation to ensure that users input correct email addresses when sharing notes. Furthermore, toast messages have been added to provide feedback to the user. This covers the scenarios where the email input does not match the required format and when a user attempts to submit a form with a yet to be added email in the field input.